Module creation. Doesn't run yet though
